fix: stop making a previous generation a precondition for booting
The `gen2` boundary was introduced to end the coupling to the old generation, and
then the code built a new dependency on it: a legacy tree this build happened to
be able to read refused startup with `legacy_adoption_required` until an operator
ran `backend store-adopt`. A fresh install with any prior history therefore began
with a hard boot failure and a migration step. That is the worst possible first
experience, and it was protecting nothing.

Nothing in the flavor root justified it. `gen2/data` holds `store`, the KB runtime
indexes, `engines`, and `equipment` — the journal plus rebuildable consumers, per
the "rebuildable consumers, not authority" principle. Everything a user authors is
outside it: the KB corpus at `~/.coral/kb`, project plans and memos under
`~/.coral/projects`, and `~/.coral/exports`. The only thing left behind is past job
history, which is exactly what cutting a generation means.

The sibling state already had the right behaviour: `legacy-foreign` leaves the old
bytes alone and initializes empty generated state. A matching fingerprint was
never a reason to migrate instead — it only told us an import was *possible*. The
two states collapse into one `legacy-ignored`: legacy bytes are left untouched, the
stored version is reported for the notice, and this generation initializes its own
state. Whether this build could read the old store no longer changes whether it
boots.

Adoption is deleted rather than automated. Automating it would have kept the
coupling and merely hidden it, and it cannot survive the cut anyway: startup now
creates the generated target, while adoption renames the legacy root *onto* that
target and refuses when it already exists (`adoptLegacyStore` guards this
explicitly). A retained command would be dead on arrival, so
`legacy-store-adoption.ts`, the `store-adopt` CLI surface, and the four
`legacy_adoption_*` codes go, guarded by an invariant so the path stays deleted.

Found by tracing the readiness branch points through the code graph: exactly three
consume it, and `adoptLegacyStore` had a single caller — the CLI command itself.
The refusal was the only thing tying startup to the old generation. One graph edge
claimed `store/db.ts` used the adoption module; reading it showed a same-named
local variable, not an import.
