+++ src/cli/classify.ts
@@ -116,7 +116,6 @@ export const commandContainerPaths = new Set<string>([
 export const commandClassExemptions = {
   'backend status': 'local operational health probe',
   'backend shutdown': 'local operational drain request',
-  'backend store-adopt': 'operator-only local same-generation store adoption',
   'backend kb-commit quarantine': 'operator-only local KB evidence quarantine',
   'backend store-reset list': 'local operational incident inspection',
   'backend store-reset report': 'local operational incident inspection',
